requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
python-magic==0.4.27
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.3.0
langchain==0.0.350
email-validator==2.1.0
aiofiles==23.2.1
jinja2==3.1.2